+++ lib/modules/cache/cache_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: cache
-// Updated at: 2026-06-03 10:20:47
-// Commit: docs(cache): update readme and API contracts for cache
+// Updated at: 2026-06-03 12:46:28
+// Commit: feat(cache): add bi-directional event bus handler for cache
 
 class cacheComponentState {
-  final String commitRef = '1684';
-  final String lastMessage = 'docs(cache): update readme and API contracts for cache';
-  final DateTime timestamp = DateTime.parse('2026-06-03T10:20:47+05:00');
+  final String commitRef = '1692';
+  final String lastMessage = 'feat(cache): add bi-directional event bus handler for cache';
+  final DateTime timestamp = DateTime.parse('2026-06-03T12:46:28+05:00');
 }